transcription-service/requirements.txt
openai-whisper
flask>=3.0.0
flask-cors>=4.0.0
